refactor: rename InfoBox component to DetailsItem
