+++ src/popup/router/pages/TransactionDetails.vue
@@ -6,7 +6,7 @@
     <div class="content">
       <div class="visual-overview"><!-- TODO --></div>
       <div class="data-grid">
-        <InfoBox
+        <DetailsItem
           v-if="tipUrl"
           :label="$t('pages.transactionDetails.tipUrl')"
           class="span-2-columns"
@@ -14,8 +14,8 @@
         >
           <CopyButton slot="label" :value="tipUrl" message="URL copied" />
           <LinkButton slot="value" :to="tipUrl">{{ tipUrl }}</LinkButton>
-        </InfoBox>
-        <InfoBox
+        </DetailsItem>
+        <DetailsItem
           :value="hash"
           :label="$t('pages.transactionDetails.hash')"
           class="span-2-columns"
@@ -23,51 +23,51 @@
           small
         >
           <CopyButton slot="label" :value="hash" message="Hash copied" />
-        </InfoBox>
-        <InfoBox
+        </DetailsItem>
+        <DetailsItem
           v-if="microTime"
           :value="microTime | formatDate"
           :secondary="microTime | formatTime"
           :label="$t('pages.transactionDetails.timestamp')"
           data-cy="timestamp"
         />
-        <InfoBox
+        <DetailsItem
           v-if="blockHeight && blockHeight > 0"
           :value="blockHeight"
           :label="$t('pages.transactionDetails.blockHeight')"
           data-cy="block-height"
         />
-        <InfoBox
+        <DetailsItem
           v-if="tx.gasUsed"
           :value="tx.gasUsed"
           :label="$t('pages.transactionDetails.gas')"
           data-cy="gas"
         />
-        <InfoBox
+        <DetailsItem
           v-if="tx.gasPrice"
           :label="$t('pages.transactionDetails.gasPrice')"
           data-cy="gas-price"
         >
           <TokenAmount slot="value" :amount="tx.gasPrice" symbol="ættos" hideFiat />
-        </InfoBox>
-        <InfoBox :label="$t('pages.transactionDetails.amount')" data-cy="amount">
+        </DetailsItem>
+        <DetailsItem :label="$t('pages.transactionDetails.amount')" data-cy="amount">
           <TokenAmount slot="value" :amount="amount" :symbol="symbol" hideFiat />
-        </InfoBox>
-        <InfoBox
+        </DetailsItem>
+        <DetailsItem
           v-if="tx.nonce"
           :value="tx.nonce"
           :label="$t('pages.transactionDetails.nonce')"
           data-cy="nonce"
         />
-        <InfoBox
+        <DetailsItem
           v-if="tx.fee"
           :label="$t('pages.transactionDetails.fee')"
           class="span-2-columns"
           data-cy="fee"
         >
           <TokenAmount slot="value" :amount="tx.fee" symbol="ættos" hideFiat />
-        </InfoBox>
-        <InfoBox
+        </DetailsItem>
+        <DetailsItem
           v-if="pending"
           :value="$t('pages.transactionDetails.pending')"
           :label="$t('pages.transactionDetails.status')"
@@ -90,7 +90,7 @@
 import { mapGetters } from 'vuex';
 import { formatDate, formatTime } from '../../utils';
 import TokenAmount from '../components/TokenAmount';
-import InfoBox from '../components/InfoBox';
+import DetailsItem from '../components/DetailsItem';
 import LinkButton from '../components/LinkButton';
 import CopyButton from '../components/CopyButton';
 import AnimatedPending from '../../../icons/animated-pending.svg?vue-component';
@@ -100,7 +100,7 @@ export default {
   name: 'TransactionDetails',
   components: {
     TokenAmount,
-    InfoBox,
+    DetailsItem,
     LinkButton,
     CopyButton,
     AnimatedPending,
